requirements.txt
atproto
python-dotenv
google-generativeai
fasthtml
sqlite3